requirements.txt
django-admin-import==0.3.0
django
djangorestframework
django-cors-headers
django-filter
